requirements.txt
django
django-allauth
requests
termcolor
simplejson
Pillow>=4.0.0
